backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Admin console BFF — FastAPI backend for the management UI"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "runtime-common",
    "argon2-cffi>=25.1.0",
    "python-multipart>=0.0.18",
    "aiofiles>=24.1",
    "aioboto3>=13.0",
    "httpx>=0.27",
    "pydantic-settings>=2.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/backend"]

[dependency-groups]
dev = [
    "aiosqlite>=0.20",
    "respx>=0.21",
]
